+++ app/src/main/java/com/fss/everythingapp/fitness/Controllers/MMSignedInController.java
@@ -1,4 +1,4 @@
-package com.fss.everythingapp;
+package com.fss.everythingapp.fitness.Controllers;
 
 import java.io.IOException;
 
@@ -8,11 +8,10 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
 import javafx.stage.Stage;
 
 
-public class PleaseProvideControllerClassName {
+public class MMSignedInController {
 
     @FXML
     void openLogMenu(ActionEvent event) {
